+++ src/AngleSharp.Diffing.Tests/Strategies/ElementStrategies/ElementComparerTest.cs
@@ -1,9 +1,22 @@
-﻿namespace AngleSharp.Diffing.Strategies.ElementStrategies;
+﻿using AngleSharp.Diffing.Core.Diffs;
+
+namespace AngleSharp.Diffing.Strategies.ElementStrategies;
 
 public class ElementComparerTest : DiffingTestBase
 {
     public ElementComparerTest(DiffingTestFixture fixture) : base(fixture)
     {
+    }
+
+    [Theory(DisplayName = "When current result is same or skip, the current decision is returned")]
+    [MemberData(nameof(SameAndSkipCompareResult))]
+    public void Test000(CompareResult currentResult)
+    {
+        var comparison = ToComparison("<p>", "<div>");
+
+        new ElementComparer(enforceTagClosing: false)
+            .Compare(comparison, currentResult)
+            .ShouldBe(currentResult);
     }
 
     [Theory(DisplayName = "When control and test nodes have the same type and name and enforceTagClosing is false, the result is Same")]
@@ -19,23 +32,28 @@ public class ElementComparerTest : DiffingTestBase
             .ShouldBe(CompareResult.Same);
     }
 
-    [Theory(DisplayName = "When control and test nodes have the a different type and name, the result is Different")]
+    [Theory(DisplayName = "When control and test nodes have the a name, the result is Different")]
     [InlineData("<div>", "<p>", false)]
-    [InlineData("<div>", "textnode", false)]
-    [InlineData("<div>", "<!--comment-->", false)]
-    [InlineData("<!--comment-->", "textnode", false)]
-    [InlineData("<br>", "<br/>", true)]
-    [InlineData("<input>", "<input/>", true)]
     [InlineData("<div>", "<p>", true)]
-    [InlineData("<div>", "textnode", true)]
-    [InlineData("<div>", "<!--comment-->", true)]
-    [InlineData("<!--comment-->", "textnode", true)]
     public void Test002(string controlHtml, string testHtml, bool enforceTagClosing)
     {
         var comparison = ToComparison(controlHtml, testHtml);
+
         new ElementComparer(enforceTagClosing)
             .Compare(comparison, CompareResult.Unknown)
-            .ShouldBe(CompareResult.Different);
+            .ShouldBe(CompareResult.FromDiff(new ElementDiff(comparison, ElementDiffKind.Name)));
+    }
+
+    [Theory(DisplayName = "When control and test nodes have the a different closing style, the result is Different")]
+    [InlineData("<br>", "<br/>", true)]
+    [InlineData("<input>", "<input/>", true)]
+    public void Test003(string controlHtml, string testHtml, bool enforceTagClosing)
+    {
+        var comparison = ToComparison(controlHtml, testHtml);
+
+        new ElementComparer(enforceTagClosing)
+            .Compare(comparison, CompareResult.Unknown)
+            .ShouldBe(CompareResult.FromDiff(new ElementDiff(comparison, ElementDiffKind.ClosingStyle)));
     }
 
     [Theory(DisplayName = "When unknown node is used in comparison, but node name is equal, the result is Same")]