+++ src/AngleSharp.Diffing.Tests/Strategies/TextNodeStrategies/StyleSheetTextNodeComparerTest.cs
@@ -32,7 +32,10 @@ public class StyleSheetTextNodeComparerTest : TextNodeTestBase
     {
         var comparison = ToStyleComparison(@"h1{background:#000;}", @"h1{color:#000;}");
 
-        StyleSheetTextNodeComparer.Compare(comparison, CompareResult.Unknown).ShouldBe(CompareResult.Different);
+        var result = StyleSheetTextNodeComparer.Compare(comparison, CompareResult.Unknown);
+
+        result.Decision.ShouldBe(CompareDecision.Different);
+        result.Diff.ShouldBeEquivalentTo(new StylesheetDiff(comparison));
     }
 
     [Theory(DisplayName = "The comparer ignores insignificant whitespace")]