+++ src/AngleSharp.Diffing/Core/Diffs/AttrDiff.cs
@@ -1,15 +1,20 @@
 namespace AngleSharp.Diffing.Core;
 
 /// <summary>
-/// Represents an attribute difference
+/// Represents an attribute difference.
 /// </summary>
-[DebuggerDisplay("Attribute diff: Control = {Control.Path}, Test = {Test.Path}")]
-public class AttrDiff : DiffBase<AttributeComparisonSource>
+public record class AttrDiff : DiffBase<AttributeComparisonSource>
 {
+    /// <summary>
+    /// Gets the kind of the diff.
+    /// </summary>
+    public AttrDiffKind Kind { get; }
+
     /// <summary>
     /// Creates an <see cref="AttrDiff"/>.
     /// </summary>
-    public AttrDiff(in AttributeComparison comparison) : base(comparison.Control, comparison.Test, DiffTarget.Attribute)
+    public AttrDiff(in AttributeComparison comparison, AttrDiffKind kind) : base(comparison.Control, comparison.Test, DiffTarget.Attribute)
     {
+        Kind = kind;
     }
 }