+++ src/AngleSharp.Diffing/Core/Diffs/UnexpectedAttrDiff.cs
@@ -3,8 +3,7 @@ namespace AngleSharp.Diffing.Core;
 /// <summary>
 /// Represents an unexpected attribute found in the test DOM tree.
 /// </summary>
-[DebuggerDisplay("Unexpected Attribute: Test = {Test.Path}")]
-public class UnexpectedAttrDiff : UnexpectedDiffBase<AttributeComparisonSource>
+public record class UnexpectedAttrDiff : UnexpectedDiffBase<AttributeComparisonSource>
 {
     /// <summary>
     /// Creates a <see cref="UnexpectedAttrDiff"/>.