+++ src/AngleSharp.Diffing/Core/Diffs/UnexpectedNodeDiff.cs
@@ -3,8 +3,7 @@ namespace AngleSharp.Diffing.Core;
 /// <summary>
 /// Represents an unexpected node in the test DOM tree.
 /// </summary>
-[DebuggerDisplay("Unexpected {Target}: Test = {Test.Path}")]
-public class UnexpectedNodeDiff : UnexpectedDiffBase<ComparisonSource>
+public record class UnexpectedNodeDiff : UnexpectedDiffBase<ComparisonSource>
 {
     /// <summary>
     /// Creates a <see cref="UnexpectedNodeDiff"/>.